+++ server/src/controllers/category.controller.js
@@ -2,11 +2,15 @@ import { Category } from "../models/category.model.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 import { removeSingleImg, uploadSingleImg } from "../utils/cloudinary.js";
 import { ApiError } from "../utils/ApiError.js";
+import { ApiResponse } from "../utils/ApiResponse.js";
+import { isValidObjectId } from "mongoose";
 
 const allCategory = asyncHandler(async (req, res, next) => {
     try {
-        const allData = await Category.find();
-        res.status(200).json(allData);
+        const categories = await Category.find();
+        res.status(200).json(
+            new ApiResponse(200, categories, "get all categories successfully")
+        );
     } catch (error) {
         next(error);
     }
@@ -32,8 +36,13 @@ const createCategory = asyncHandler(async (req, res, next) => {
             description,
             thumbnail: thumbnailPath,
         });
+        if (!category) {
+            throw new ApiError(401, "create new category failed");
+        }
 
-        return res.status(200).json(category);
+        return res
+            .status(200)
+            .json(new ApiResponse(200, {}, "create new category successfully"));
     } catch (error) {
         next(error);
     }
@@ -44,10 +53,19 @@ const updateCategory = asyncHandler(async (req, res, next) => {
     const thumbnail = req.file;
     const { name, description } = req.body;
     try {
-        const category = await Category.findOne({ _id: req.params.categoryId });
+        const { categoryId } = req.params;
+
+        if (!isValidObjectId(categoryId)) {
+            throw new ApiError(401, "this category ID is not valid");
+        }
+
+        const category = await Category.findOne({ _id: categoryId });
+        if (!category) {
+            throw new ApiError(401, "this category not found on database");
+        }
 
         if (thumbnail?.path) {
-            const thumbnailPath = await uploadSingleImg(thumbnail?.path);
+            const thumbnailPath = await uploadSingleImg(thumbnail.path);
             if (thumbnailPath) {
                 await removeSingleImg(category.thumbnail);
                 category.thumbnail = thumbnailPath;
@@ -63,12 +81,14 @@ const updateCategory = asyncHandler(async (req, res, next) => {
         }
 
         if (!category) {
-            throw new ApiError(401, "not update category failed");
+            throw new ApiError(401, "category update failed on db");
         }
 
         await category.save();
 
-        return res.status(200).json(category);
+        return res
+            .status(200)
+            .json(new ApiResponse(200, {}, "update category successfully"));
     } catch (error) {
         next(error);
     }
@@ -76,20 +96,24 @@ const updateCategory = asyncHandler(async (req, res, next) => {
 
 const deleteCategory = asyncHandler(async (req, res, next) => {
     try {
-        const deleted = await Category.findOneAndDelete({
-            _id: req.params.categoryId,
-        });
+        const { categoryId } = req.params;
+        if (!isValidObjectId(categoryId)) {
+            throw new ApiError(401, "this category ID is not valid");
+        }
 
-        await removeSingleImg(deleted.thumbnail);
+        const deleted = await Category.findOneAndDelete({
+            _id: categoryId,
+        });
 
         if (!deleted) {
             throw new ApiError(404, "category not deleted on database");
         }
 
-        return res.status(200).json({
-            message: "category deleted successfully",
-            statusCode: 200,
-        });
+        await removeSingleImg(deleted.thumbnail);
+
+        return res
+            .status(200)
+            .json(new ApiResponse(200, {}, "category deleted successfully"));
     } catch (error) {
         next(error);
     }
